pyproject.toml
[project]
name = "bluesky-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.55",
    "python-decouple>=3.8",
    "sqlite-utils>=3.38",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.4",
]
